install-deps: install ssh to fix issue with connectivity
If you are going to install a slew of packages after updating
the source list, chances are you may be updating a library which
a distribution somehow has not yet annotated as requiring sshd to
be updated as well, and when doing the install of packages ssh
will stop working.

I've ran into this issue now with terraform on debian testing
on both azure and gce debian 10 images upgrading to debian testing.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>