install-deps: upgrade for rolling distributions
On the debian front we rely on debian tests. On the suse side
we have a knob which disables upgrading packages for SLE releases,
and on fedora we are using the latest packages.

The previous commit reported an issue when installing a slew of
preferred packages: sometimes you may upgrade a library and that
may sometimes break ssh if a library was updated which had a library
ssh depended on. At least on debian testing I am seeing this issue.

Try to avoid *any* possiblity for this to happen as we are using
devconfig when provisioning, and if we break ssh, well that breaks
everything.

Instead take the safe side of always doing the upgrade. If you
know you should not or cannot do this on your release you can
add rules for this to special case it.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>