devconfig: split install deps semantics
We want to be able to easily test kdevops, but right now we have
tons of things which the devconfig does by default if enabled
which takes up a lot of time.

To help reduce time to test with bare bones things, we add
new semantics to split up what kdevops does to install packages
into a few items:

  * refresh repos
  * update system
  * install some basic kernel development tools

The refresh repos aspect just updates your repositories, so that
if you later do install something, it will be using a fresh copy
of the repository's data. On systems where you have to register
your system, such as SLES, updating your repositories only makes
sense if you are able to register your system, and so refreshing
repos implies trying to also register your system if possible.

Updating the system takes the most amount of time. It will update
all of your packages.

Installing some basic kernel development tools is the last step.
Note that we have some workflows already, where package dependencies
can be spelled out. And so, we must not abuse the devconfig role with
by putting in everyone's preferred package there. In fact, we should
consider later trimming this. We want to strive for generic kernel
development tools. Note that we already have a bootlinux role where
we install package dependencies to compile a kernel, and install it.
So there is some current overlap between the two. Maybe we should
trim the package list we have from devconfig some more.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>