SUSE: fix installation of fio on SLES12-SP1 when registering a system
We currently always install fio when a system is being registered.
SLES12-SP1 has the fio package, however its part of a special
product, ses/2.1/x86_64, which you must register for, and this
*also* requires another special registration code for it.

Instead of adding support to register for custom things, we leave
that for the target workflow subsystem to resolve. The devconfig
ansible role only cares to try to install general kernel development
tools, *when possible*, if they are not available, and if it requires
too much work, we lave that to other workflows to resolve. For
instance, once kdevops gets support for blktests, that work can be
added there.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>